Upload NoteMaster10000
The Note-Master 10000 is a multi-purpose note taking, organizing, and storing tool that also helps its users make new and unexpected conceptual connections between ideas. The impetus behind the Note-Master 10000 was my wish to have note taking software that allowed me to enter notes and ‘tag’ them with keywords, and then look-up all the notes I had tagged with a given keyword. 

The Note-Master achieves that functionality by storing notes as sets of strings in a dictionary. The keyword is a saved as a dictionary key, and when a new note (string) is tagged with a keyword, it is added to set of strings under the corresponding key. Besides allowing users to create and tag notes, Note-Master 10000 allows users to delete notes and keywords from the dictionary. 

All changes made to the notes dictionary are saved at the time of modification by ‘pickling’ them into an external .dat file. This allows users to maintain and return to their dictionary across uses of the program. 

Finally, the Note-Master 10000 helps spark new ideas by drawing connections between the notes filed in its dictionary. The Combobulator function randomly selects two notes with at least one word in common (common words are disregarded, to avoid making trivial connections). It then prints the two notes and offers a list of common words with the suggestion that they may indicate ‘points of conceptual convergence’ for the user to elaborate on.